restuctured code into multiple .py files corresponding to classes
