+++ requirements_dev.txt
@@ -9,3 +9,4 @@ Sphinx==1.8.1
 twine==1.12.1
 pytest==3.8.2
 pytest-runner==4.2
+kivy>=1.0.6
