+++ sarcomereannotation/__init__.py
@@ -11,4 +11,5 @@ def get_module_version():
     return __version__
 
 
-from .example import Example  # noqa: F401
+from .SarcomereLines import SarcomereLines
+from .Picture import Picture
