--- sarcomereannotation/bin/my_example.py
@@ -1,76 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-"""
-This sample script will get deployed in the bin directory of the
-users' virtualenv when the parent module is installed using pip.
-"""
-
-import sys
-import argparse
-import logging
-import traceback
-from sarcomereannotation import get_module_version, Example
-
-###############################################################################
-
-log = logging.getLogger()
-logging.basicConfig(level=logging.INFO,
-                    format='[%(levelname)4s:%(lineno)4s %(asctime)s] %(message)s')
-
-###############################################################################
-
-
-class Args(argparse.Namespace):
-
-    DEFAULT_FIRST = 10
-    DEFAULT_SECOND = 20
-
-    def __init__(self):
-        # Arguments that could be passed in through the command line
-        self.first = self.DEFAULT_FIRST
-        self.second = self.DEFAULT_SECOND
-        self.debug = False
-        #
-        self.__parse()
-
-    def __parse(self):
-        p = argparse.ArgumentParser(prog='run_exmaple',
-                                    description='A simple example of a bin script')
-        p.add_argument('-v', '--version', action='version', version='%(prog)s ' + get_module_version())
-        p.add_argument('-f', '--first', action='store', dest='first', type=int, default=self.first,
-                       help='The first argument value')
-        p.add_argument('-s', '--second', action='store', dest='second', type=int, default=self.second,
-                       help='The first argument value')
-        p.add_argument('--debug', action='store_true', dest='debug', help=argparse.SUPPRESS)
-        p.parse_args(namespace=self)
-
-
-###############################################################################
-
-def main():
-    try:
-        args = Args()
-        dbg = args.debug
-
-        # Do your work here - preferably in a class or function,
-        # passing in your args. E.g.
-        exe = Example(args.first)
-        exe.update_value(args.second)
-        print("First : {}\nSecond: {}".format(exe.get_value(), exe.get_previous_value()))
-
-    except Exception as e:
-        log.error("=============================================")
-        if dbg:
-            log.error("\n\n" + traceback.format_exc())
-            log.error("=============================================")
-        log.error("\n\n" + str(e) + "\n")
-        log.error("=============================================")
-        sys.exit(1)
-
-
-###############################################################################
-# Allow caller to directly run this module (usually in development scenarios)
-
-if __name__ == '__main__':
-    main()
